+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.23.1] - Unpublished
+
+### Added
+
+- Object options test individual flags
+
+### Fixed
+
+- Object fetch options [bug](https://github.com/FrankC01/pysui/issues/110)
+
+### Changed
+
+### Removed
+
 ## [0.23.0] - 2023-05-23
 
 ### Added